feat: Add getWatchHistory endpoint
This commit adds the getWatchHistory endpoint to the user.controller.js file. The endpoint retrieves the watch history of a user by aggregating data from the User and Video collections. It includes a lookup to get the owner information for each video in the watch history. The endpoint returns the watch history along with a success message. This feature enhances the functionality of the application by allowing users to view their watch history.