feat: Add getVideoDuration function to calculate video duration
This commit adds the getVideoDuration function to the duration.js file in the utils folder. The function uses the fluent-ffmpeg library to retrieve the duration of a video file. It takes the video path as a parameter and returns a promise that resolves with the duration in seconds. This addition enhances the functionality of the application by providing a convenient way to calculate the duration of videos.